requirements.txt
selenium
pytest
pyyaml
faker

